requirements.txt
attrs==18.1.0
backports.functools-lru-cache==1.5
beautifulsoup4==4.6.3
certifi==2018.8.13
chardet==3.0.4
ConfigArgParse==0.13.0
configparser==3.5.0
coursera-dl==0.11.4
cycler==0.10.0
decorator==4.2.1
Django==1.11.15
djangorestframework==3.8.2
entrypoints==0.2.3
idna==2.7
keyring==13.2.1
kiwisolver==1.0.1
matplotlib==2.2.2
networkx==2.1
numpy==1.14.2
pyasn1==0.4.4
pyparsing==2.2.0
python-dateutil==2.7.2
pytz==2018.3
pywin32-ctypes==0.1.2
requests==2.19.1
six==1.11.0
urllib3==1.23
